+++ src/App.js
@@ -1,7 +1,8 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { Link, Switch, Route } from 'react-router-dom';
+import { Link, Switch, Route, withRouter } from 'react-router-dom';
 
+// Tori contracts
 import ToriToken from '../build/contracts/ToriToken.json';
 import ToriVisit from '../build/contracts/ToriVisit.json';
 // Accessories contract
@@ -13,21 +14,33 @@ import ClothCushion from '../build/contracts/ClothCushion.json';
 
 import getWeb3 from './utils/getWeb3';
 
-import MyToriDisplay from './MyToriDisplay.js';
-import Inventory from './Inventory.js';
-import Trade from './Trade.js';
+
+import Info from './Info/Info.js';
+import MyTori from './MyTori/MyTori.js';
+import EditRoom from './MyTori/EditRoom.js';
+import Nursery from './Nursery/Nursery.js';
+import Fuse from './Nursery/Fuse.js';
+import BreedHome from './Nursery/BreedHome.js';
+import Breed from './Nursery/Breed.js';
+import Inventory from './Inventory/Inventory.js';
+import OtherToris from './Explore/OtherToris.js';
+import ToriDetails from './Explore/ToriDetails.js';
+import Market from './Marketplace/Market.js';
+import Confirmation from './Components/Confirmation.js';
+
 
 import './css/oswald.css'
 import './css/open-sans.css'
 import './css/pure-min.css'
 import './App.css'
 
-
+import { withStyles } from '@material-ui/core/styles';
 import AppBar from '@material-ui/core/AppBar';
 import Tabs from '@material-ui/core/Tabs';
 import Tab from '@material-ui/core/Tab';
 import Typography from '@material-ui/core/Typography';
 import Toolbar from '@material-ui/core/Toolbar';
+import Snackbar from '@material-ui/core/Snackbar';
 
 function TabContainer(props) {
   return (
@@ -41,6 +54,18 @@ TabContainer.propTypes = {
   children: PropTypes.node.isRequired,
 };
 
+const styles = theme => ({
+  tab: {
+    position: 'absolute',
+    right: 0,
+    marginRight: 20,
+  },
+  banner: {
+    backgroundColor: theme.palette.secondary.dark,
+    height: 20,
+    width: `100%`,
+  },
+});
 
 class App extends Component {
 
@@ -50,19 +75,31 @@ class App extends Component {
     toriVisit: PropTypes.object,
     accContracts: PropTypes.array,
     userAccount: PropTypes.string,
+    onMessage: PropTypes.func,
   }
 
   constructor(props) {
     super(props)
 
+    let loc = this.props.history.location.pathname;
+    loc = loc.split('/')[1];
+
+    const locMode = {'': 4, 'mytoris': 0, 'nursery': 1, 'inventory': 2, 'explore': 3, 'market': 4}
+    let currentMode = locMode[loc];
+
     this.state = {
       storageValue: 0,
       web3: null,
-      mode: 0,
+      mode: currentMode === undefined ? locMode[''] : currentMode,
       accessoriesTokenInstances: [],
+      accNum: 100,
+      openSnackBar: false,
     }
 
+    // Function BINDS
     this.switchDisplay = this.switchDisplay.bind(this);
+    this.handleCloseSnackBar = this.handleCloseSnackBar.bind(this);
+    this.handleMessage = this.handleMessage.bind(this);
   }
 
   getChildContext() {
@@ -72,10 +109,11 @@ class App extends Component {
       toriVisit: this.state.toriVisitInstance,
       accContracts: this.state.accessoriesTokenInstances,
       userAccount: this.state.userAccount,
+      onMessage: this.handleMessage
     };
   }
 
-  componentWillMount() {
+  async componentDidMount() {
     // Get network provider and web3 instance.
     // See utils/getWeb3 for more info.
 
@@ -89,8 +127,25 @@ class App extends Component {
       this.instantiateContract();
     })
     .catch(() => {
-      console.log('Error finding web3.')
-    })
+      this.onMessage('Error finding web3.');
+    });
+
+    // Periodically check if the metamask account has changed
+    setInterval( async () => {
+      if (this.state.web3 !== undefined) {
+        this.state.web3.eth.getAccounts((error, accounts) => {
+          if (this.state.userAccount !== accounts[0]) {
+            // Redirect ...
+            this.setState({
+              userAccount: accounts[0],
+              mode: 4,
+            }, () => {
+              if (this.props.history.location.pathname !== "/") this.props.history.push('/');
+            });
+          }
+        });
+      }
+    }, 1000);
   }
 
   instantiateContract() {
@@ -114,6 +169,9 @@ class App extends Component {
     const cc = contract(ClothCushion);
 
     let accessories = [wd, wc, ws, wb, cc];
+    this.setState({
+      accNum: accessories.length,
+    });
     accessories.forEach((c) => c.setProvider(this.state.web3.currentProvider));
 
     // Get accounts.
@@ -131,14 +189,15 @@ class App extends Component {
       });
 
       // Tori Accessories
-      accessories.forEach((c) => {
-        c.deployed().then((instance) => {
+      Promise.all(accessories.map((c) => c.deployed()))
+      .then((instances) => {
+        instances.forEach((instance) =>
           this.setState({
             accessoriesTokenInstances: this.state.accessoriesTokenInstances.concat(instance),
-          });
-        });
-      });
-    })
+          })
+        );
+      })
+    });
   }
 
   switchDisplay(e, mode) {
@@ -147,8 +206,20 @@ class App extends Component {
     });
   }
 
+  handleCloseSnackBar() {
+    this.setState({
+      openSnackBar: false,
+    });
+  }
+
+  handleMessage(message) {
+    this.setState({
+      openSnackBar: true,
+      snackBarMessage: message,
+    });
+  }
+
   render() {
-    // this.state.currentDisplay
     return (
       <div className="App">
         <AppBar position="static">
@@ -156,25 +227,49 @@ class App extends Component {
             <Typography variant="title" color="inherit">
               Cribtori
             </Typography>
+            <Tabs value={this.state.mode}
+                  onChange={this.switchDisplay}
+                  className={this.props.classes.tab}>
+              <Tab label="My Toris" component={Link} to={'/mytoris'} />
+              <Tab label="Nursery" component={Link} to={'/nursery'} />
+              <Tab label="Inventory" component={Link} to={'/inventory'} />
+              <Tab label="Explore" component={Link} to={'/explore'} />
+              <Tab label="Marketplace" component={Link} to={'/market'} />
+              <Tab label="Info" component={Link} to={'/'} />
+            </Tabs>
           </Toolbar>
-          <Tabs value={this.state.mode} onChange={this.switchDisplay} centered>
-            <Tab label="My Toris" component={Link} to={'/'} />
-            <Tab label="Inventories" component={Link} to={'/inventory'} />
-            <Tab label="Other Toris" component={Link} to={'/others'} />
-            <Tab label="Yard Sale" component={Link} to={'/trade'} />
-          </Tabs>
         </AppBar>
-          {this.state.toriTokenInstance &&
-            <Switch>
-              <Route exact path='/' component={MyToriDisplay} />
-              <Route path='/inventory' component={Inventory} />
-              <Route path='/others' component={MyToriDisplay} />
-              <Route path='/trade' component={Trade} />
-            </Switch>
-          }
+        <div className={this.props.classes.banner}>
+        </div>
+        {(this.state.accessoriesTokenInstances.length === this.state.accNum) &&
+         (this.state.userAccount) &&
+          <Switch>
+            <Route exact path='/' component={Info} />
+            <Route exact path='/mytoris' component={MyTori} />
+            <Route exact path='/mytoris/edit' component={EditRoom} />
+            <Route exact path='/nursery' component={Nursery} />
+            <Route exact path='/nursery/fuse' component={Fuse} />
+            <Route exact path='/nursery/breed' component={BreedHome} />
+            <Route exact path='/nursery/breed/:id' component={Breed} />
+            <Route exact path='/inventory' component={Inventory} />
+            <Route exact path='/explore' component={OtherToris} />
+            <Route exact path='/explore/:id' component={ToriDetails} />
+            <Route exact path='/market' component={Market} />
+            <Route exact path='/confirmation' component={Confirmation} />
+          </Switch>
+        }
+        <Snackbar
+          anchorOrigin={{ vertical: 'bottom', horizontal: 'center' }}
+          open={this.state.openSnackBar}
+          onClose={this.handleCloseSnackBar}
+          snackbarcontentprops={{
+            'aria-describedby': 'message-id',
+          }}
+          message={<span id="message-id">{this.state.snackBarMessage}</span>}
+        />
       </div>
     );
   }
 }
 
-export default App
+export default withStyles(styles)(withRouter(App))
