+++ src/Components/Timer.js
@@ -1,5 +1,4 @@
-import React, { Component } from 'react'
-import PropTypes from 'prop-types';
+import React, { Component } from 'react';
 
 import { withStyles } from '@material-ui/core/styles';
