+++ src/Nursery/Nursery.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
-import PropTypes from 'prop-types';
-import { Link, Switch, Route, withRouter } from 'react-router-dom';
+import { withRouter } from 'react-router-dom';
 
 import { withStyles } from '@material-ui/core/styles';
 import Grid from '@material-ui/core/Grid';
@@ -55,9 +54,9 @@ const styles = theme => ({
 });
 
 class Nursery extends Component {
-  constructor(props) {
+  constructor(props, context) {
     super(props);
-
+    this.context = context;
   }
 
   render() {
