+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { HashRouter } from 'react-router-dom';
 
-import App from './App2';
+import App from './App';
 
 import { MuiThemeProvider, createMuiTheme } from '@material-ui/core/styles';
